move: enums 경로 이동 (27)
