+++ src/main/java/modelly/modelly_be/domain/recruitment/dto/response/RecruitmentListResponseDto.java
@@ -1,6 +1,6 @@
 package modelly.modelly_be.domain.recruitment.dto.response;
 
-import modelly.modelly_be.domain.recruitment.entity.SubCategory;
+import modelly.modelly_be.domain.recruitment.entity.enums.SubCategory;
 import modelly.modelly_be.global.entity.Category;
 
 import java.time.LocalDateTime;
