+++ src/main/java/modelly/modelly_be/domain/recruitment/entity/enums/SubCategory.java
@@ -1,4 +1,4 @@
-package modelly.modelly_be.domain.recruitment.entity;
+package modelly.modelly_be.domain.recruitment.entity.enums;
 
 public enum SubCategory {
     //HAIR
