+++ src/main/java/modelly/modelly_be/domain/user/dto/response/DesignerListResponseDto.java
@@ -1,6 +1,5 @@
 package modelly.modelly_be.domain.user.dto.response;
 
-import modelly.modelly_be.domain.recruitment.entity.SubCategory;
 import modelly.modelly_be.global.entity.Category;
 
 import java.time.LocalDateTime;
